extract TrustOptions into its own file
